Added constructors for ConfigProvider
We will have 2 constructors. One that will allow the user to specify the path to the directory to the config json file and its name (e.g. appsettings.json). The other allows the user to load the config file by only specifying its name. In this case, we will default to the project's directory.