Added GetConfigurationMethod
Next we added a private method to get our configuration instance.
To use the setBasePath method we needed to install the nuget Microsoft.Extensions.Configuration.Binder.
To use the AddJsonFile method we needed to install the nuget Microsoft.Extensions.Configuration.Json.

Before calling Build, we have a configuration builder. Once we call the method Build, it creates an IConfiguration instance with the keys and values from a set of defined sources. For now, that is only the file that we will pass to it.