Implemented MapAkkaConfig
So this implementation is simple.  We get the section "akka" from the configuration and use the built-in Get method to map the key-value pairs to the type T. We do not know anything about the Type T.
It is the job the type T to correctly define how the mapping should work.