Implemented and fixed return type of GetAkkaConfig.
I made a mistake in the previous commit. The return Type of GetAkkaConfig should actually be Config from Akka.Configuration. This is because we want to return the finished Config instance.

In the GetAkkaConfig method we first call MapAkkaConfig to get an instance of type T. Afterwards, we create an anonymous object with the field akka. We basically wrap the config in a new object to include the key "akka". Otherwise, this key will be missing from the config instance.
We then use the configuration factory to build a valid Akka.Configuration.Config instance from our anonymous object.