Added some demo code
We want to create a router, that is defined in the configuration. To use a router defined in the configuration, we need to give it the appropriate name (look at appsettings.json).
We then send 5 messages to the router to demonstrate that this all works.
We also use FromConfig.Instance when defining the router.

Note: this will not work yet.