Added MyAkkaConfig class for configuration
If we take a look at the appsettings.json file, we will notice that it basically consists of nested json objects. Therefore, we need to reflect that structure in our configuration instance.

That means that the field, for example actor, needs to be mapped to an object that contains the relevant fields. Another example, router needs to be mapped to an object that has the fields for nr-of-instances and router.

For simplicity we define these config sections as nested classes inside our MyAkkaConfig. We could have defined each of these nested classes outside or in a separate file.