added properties to MyAkkaConfig class and subclasses
So, the names of the properties defined in our class need to match the names of the fields in appsettings.json. Look at appsettings.json. Under the key "deployment" there is an object whose key is "myrouter". Therefore, our property needs to be called MyRouter. The same is true for the fields in our router object in settings. The properties in our json config file are called router and NrOfInstances. And the same names need to be used in our RouterConfig class.