Added JsonProperty attributes
Finally, we need to define how our properties will be serialized/deserialized into json properties.
We use the JsonProperty attributes and the PropertyName property. The JsonProperty attribute is defined in Newtonsoft.Json.
The names of the properties defined by PropertyName must match how the fields would be defined in HOCON.